requirements.txt
numpy
Pillow
scikit-learn
scipy
timm
tqdm
wget
tifffile
scikit-image
kornia
imageio
tensorboard
opencv-python
open3d
setuptools==59.5.0;